Enable CORS and change field type in model
See in
https://docs.djangoproject.com/en/4.1/ref/models/fields/
https://pypi.org/project/django-cors-headers/

+++ transaction/models.py
@@ -7,7 +7,7 @@ class Transaction(models.Model):
     title = models.CharField(max_length=300)
     type = models.CharField(max_length=300)
     category = models.CharField(max_length=300)
-    amount = models.DecimalField(decimal_places=2, max_digits=10)
+    amount = models.IntegerField()
     date = models.DateTimeField()
     createdAt = models.DateTimeField(auto_now=True)
